Refactor: NetworkProvider
The reason behind adding state management for NetworkProvider was a need for an information that frontend is "connecting" to a new network. First idea was to add new parameter to NetworkProvider, but it would make it unreadable.

This way NetworkProvider has separate states (connecting, connected, disconnected) for every event (ConnectTo, SetUp, Disconnect) taking place in app.