Bugfix: Remove public key from SaifuWallet
SaifuWallet is a type of wallet that does not contain a private key. It can be created from bech32 address. Bech32 contains hrp name & wallet address bytes. The address is public key checksum. We can't decode the hash, so in this case we can't create the public key from bech32