Bugfix: Update NetworkUnknown to NetworkOffline
In some cases, connected network was always displayed as NetworkUnknownModel. The reason of this was when we were trying to connect to server with "http" prefix and request failed, NetworkModuleService made a second attempt with "https" prefix. If that also failed, NetworkModuleService returned NetworkOfflineModel with an url containing latest prefix (https). NetworkModuleBloc treats this case as a new server and doesn't emit new connected state. This was causing frontend to continue showing it as NetworkUnknownModel.

List of changes:
- compare servers URI host name instead of full URI