Feature: Loading page
Created Loading Page, that shows up when the application is launched.
This feature is a visual representation of operations handled by
NetworkModuleInitEvent.

Currently, when network is connecting, app displays page with relevant
information. When the status is downloaded, and it has healthy state,
user is redirected to requested page. Otherwise, if network is unhealthy
or offline app is redirecting to the page with network list where user
must specify online server to connect to.

List of changes:
- renamed pages_wrapper.dart into menu_wrapper.dart and moved it to
lib/pages/menu directory
- created new pages_wrapper.dart wrapping entire application to manage
loading_page.dart visibility
- created loading_list_page_cubit.dart and loading_page.dart
- created network_list_page.dart, containing list of available networks
(like in network_drawer_page.dart)