Refactor: Transaction fee based on transaction type
Previously, transaction fee was taken from the QueryNetworkProperties
endpoint - from minTxFee value.
It was implemented this way because the endpoint that returns fee per
transaction was not yet available.
Now transactions use the QueryExecutionFees endpoint to retrieve fee
values.

List of changes:
- created tx_form_initializer_cubit.dart and tx_form_builder_cubit.dart
- containing functionality extracted from tx_send_form_cubit.dart
- removed tx_send_form_cubit.dart
- created interx_msg_types.dart containing map with transaction type as
key and interx transaction name as value. It’s needed to link
transaction types used in miro with those used in interx.