Feature: Added optional connection to proxy server
After deploying miro application on IPFS, ERR_SSL_PROTOCOL_ERROR
occurred for most servers. This happened because of the differences
between the protocols used in frontend-server communication. To use the
application correctly, both - interx and frontend must use the same
HTTPS (SSL) or HTTP protocol.

List of changes:
- added proxy support for servers using HTTP. If frontend uses HTTPS,
all requests to a server that have HTTP are forwarded to proxy with
HTTPS enabled. However, if frontend uses HTTP - requests are routed
directly to the selected server.
- changed default protocol to HTTPS from HTTP (for custom networks not
containing it)
- renamed methods in network_utils.dart - parseUrl() ->
parseUrlToInterxUri(), _parseToUriWithSchema() -> parseNoSchemeToHTTPS()
- adjusted tests after changes