Bugfix: Url path on ipfs
After deploying the application on IPFS, it wasn't possible to open the
application from an url other than index.html. This happened because
IPFS doesn't host application from root url, which prevents the correct
loading of assets.

List of changes:
- removed setPathUrlStrategy() from main.dart. Now the root url is
determined by the '#'
- changed queryParameter getter and setter in
browser_url_controller.dart to extractQueryParameters() and
replaceQueryParameters() methods (with new functionality). After
removing setPathUrlStrategy, previous version didn't recognize '#' as a
valid url component. Because of this, it wasn't possible to read
queryParameters from the url.