Feature: Migration to Flutter 3.3.9
Previous version of Flutter (3.0.5) had problems related to the mobile
version:
- "confirm" action for the text field was not supported by the phone's
browser. (Tested on Android Chrome 108.0.5359.79). This bug was fixed in
the latest version of Flutter
- bug related with ticket “toast-error” - multiple toast calls causing
them to overlap. This was a bug in the "fluttertoast" package and was
fixed by updating the package to the latest version (8.1.1)

List of changes:
- updated Flutter to version 3.3.9
- updated packages to the latest supported versions
- resolved linter warnings related with
“no_leading_underscores_for_local_identifiers”
(broadcast_service_test.dart - unit and integration)
- updated README.md and deploy.sh