Bugfix: Delimiter in public address short version (#12)
Previously, short version of bech32 address always used the underscore as a delimiter. This value is used in widgets related with account (account_tile.dart, signed_in_account_button.dart) and as keyfile name. Delimiter for account-related widgets should be three dots and for the keyfile name should be an underscore.

List of changes:
- changed "bech32Shortcut" getter in "wallet_address.dart" to "buildBech32AddressShort()" method with required “delimiter” parameter