Refactor: CreateWalletPageCubit (#22)
Previously create_wallet_page.dart contained the wallet generation logic directly in the widget. This was inconsistent with the rest of the application. Additionally, it was noticed that sections in create_wallet_page.dart didn't collapse automatically after pressing "Generate new address" button. For this reason, instead of creating a bugfix for the old code, the page was completely rebuilt from the ground up.

List of changes:
- created “create_wallet_page_cubit.dart” to handle wallet creation for “create_wallet_page.dart”
- changed “download_keyfile_section.dart” and “download_keyfile_section_controller.dart” due to new cubit
- changed “controllable_expansion_tile.dart”, “kira_expansion_tile.dart”, “kira_expansion_tile_controller.dart” to fix auto-collapse when generating new address in "create_wallet_page.dart"
- renamed "toggle" into "invertVisibility" throughout the project, due to its non-descriptive naming
- renamed "validate" into "reloadErrorMessage" in "kira_text_field_controller.dart", due to its non-descriptive naming