Refactor: Scroll views
Dashboard Page was the only page that used a custom scroll widget FilledScrollView instead of Slivers. Because this widget did not provide any special features or value and the assumptions regarding the application have changed, it was decided to replace it with a generic Sliver.

Additionally, if list widgets were empty, they were still visible as unnecessarily big containers which would not scale along with the screen.

List of changes:
- deleted filled_scroll_view.dart
- replaced FilledScrollView with CustomScrollView and SliverToBoxAdapter in DashboardPage
- changed min. height from 500 to 50 in list_no_results_widget.dart, list_error_widget.dart and list_loading_widget.dart
- the above list widgets will now scale down when displayed on small screens instead of being a blank scrollable container