Experimental: Network Visualizer
The Network Visualizer is a feature that displays the map of peers of connected network. Objective of the visualizer is to allow new network participants discover available nodes and make rational choices when connecting to ensure that network remains decentralized. The data is also displayed in the table below the map. For more details, see the documentation.

This feature introduces a new page accessed from nav menu, to showcase nodes present in each individual country both as a map and a table.

List of changes:
- created new method in api_repository.dart to fetch QueryP2P
- created assets_repository.dart to fetch data from country_codes_lat_long.json and other files from assets in the future
- created query_visualizer_service.dart and country_lat_long_service.dart
- created DTOs for P2P and country_codes_lat_long.json response
- created VisualizerModel and CountryLatLongModel
- created mocked responses for P2P endpoint
- added FlutterMap dependency which provides free map widget for flutter
- added nav tile and provided new icon from Flutter Icons
- enabled dropdown scrolling if list of filters is longer than 4 items
- added json file with latitude and longitude of every country to assets, because interx provides only country codes
- added copy button for easier ip access
- added optional parameters for aligning filter chips, to improve design flexibility on different pages and devices
- added tests for Visualizer: Model, Service and Filter Options