Bugfix: SearchBar & ListSearchWidget (#39)
All SearchBar widgets on currently existing lists (ValidatorsPage, BalancesPage) had a bug related to loss of state. Any update to the widget or resizing of the browser window cleared the SearchBar value. It was giving the impression of a broken list because a list was still filtered despite empty value in TextField. This happened because list_search_widget.dart is rebuilt every time when parent widget "title" from paginated_infinity_list.dart is updated. Because of this, after each rebuild, a new TextEditingController, with an empty default value, is assigned to the SearchBar.

List of changes:
- extracted declaration of the  SearchBar TextEditingController  to the topmost widget, associated with the list (ValidatorsPage, BalancesPage), to keep a single TextEditingController for the entire lifecycle of widget
- updated default value for “clear button” visibility in search_bar.dart to keep this button visible after the widget is rebuilt
- updated balances_filter_options.dart, validators_filter_options.dart due to missing lowercase pattern parsing. Previously, the SearchBar in BalancePage correctly searched for values only for lowercase letters. When entered text in uppercase letters (e.g. KEX), the results were empty.