Bugfix: List reload on network change
Recently merged "dp-feature/paginated-list" (#38) branch had a bug related to keeping the same page number, after changing the network. This means if a list page number was changed and later the network was switched, the page number remained the same. This makes sense if the networks have the same results, however, if the network has fewer items in the list, the page numbers don't get updated or restarted, and the view responds with “No results”.

List of changes:
- created a new parameter in the ListNextPageEvent event to store reload status information
- updated the "jumpToTop" parameter in the ListNextPageEvent event handler, based on the reload status, to return to the first page after a network change