(to delete before merge) Feature: My Account Staking Tab UI
This feature introduces new tab containing list of validators, to manage staking tokens process. This is the base section for Delegation, Undelegation and Claim Rewards transactions, that will be implemented afterward. It will allow delegating and undelegating tokens to specific validators and to claim the rewards generated by the network.

List of changes:
- added Staking tab in My Account Page
- created new drawer page accessed from My Account - Staking tab, which provides staking pool details about selected Validator
- created staking_pool_drawer_cubit.dart to manage the states of staking pool data in the drawer