(to delete before merge) Feature: Delegate Tokens Transaction UI
This feature introduces new page containing transaction process for delegating tokens to a specific validator. It allows selecting the token, inputting a value, then signing the transaction and sending this data to the blockchain to delegate this token to selected validator. To keep the workflow simplicity, this branch does not yet support delegating multiple tokens.

List of changes:
- created new page staking_tx_delegate_page.dart containing implementation of transaction process for MsgDelegate
- created form widgets for MsgDelegate (staking_msg_delegate_form.dart, staking_msg_delegate_form_preview.dart)
- added buttons to initialize Delegate transaction in 3 sections: Validator Drawer Page, My Account - Staking tab and Staking Drawer Page