Bugfix: Missing Disposals
Some pages and widgets contained objects that were not disposed inside the widget. This included instantiations of controllers, blocs/cubits, notifiers, listeners or global keys. Such action progressively damaged the performance of the application. The purpose of this branch is adding missing disposals in all widgets. However, this issue is not relevant for globally accessed objects, which are supposed to remain active.

List of changes:
- added overridden "dispose()" method to classes with missing disposals
- added "close()" method to download_keyfile_section_controller.dart and kira_text_field_controller.dart to enable closing/disposing their contents
- changed all "BlocProvider<>" constructors to "BlocProvider<>.value()" to enable custom disposing and solve issues with double disposing/not disposing at all
- added condition for showLoadingOverlay notifier in a_list_bloc.dart, so it will not be used after closing the widget
- removed unnecessary TokenFormCubit from ir_tx_request_verification_page.dart, ir_tx_request_verification_form_dialog.dart and ir_msg_request_verification_form.dart. It was never used there, so instead of disposing, it was removed completely.
- removed double declaration of stream listener in a_list_bloc.dart