Feature: Identity Registrar - App integration
There were several places in the app that should display entries from Identity Registrar ("avatar" and "username" for user records, and "logo" for validator record). Instead, they displayed widgets not integrated with Identity Registrar, because it was missing during their development. The purpose of this branch is to integrate already existing widgets in the app to display entries from Identity Registrar.

List of changes:
- updated my_account_button.dart (its mobile and desktop versions) and account_header.dart to display avatar and username
- updated kira_identity_avatar.dart with a new "avatarUrl" parameter identifying the user's image avatar
- created 2 separate widgets for kira_identity_avatar.dart, so it can handle different types of avatars: image avatar (url_avatar_widget.dart) and unique gravatar (jdenticon_gravatar.dart)
- added handling empty avatar and loading functionality in kira_identity_avatar.dart
- updated validator_list_item_desktop.dart and validator_list_item_mobile.dart to provide the URL of the validator logo
- changed "social_media" key to "social" in ir_model.dart as it is official key reserved for social media