Bugfix: TokenDropdown
Recently merged branch "Feature: Transaction List Filters (#61)" introduced a bug related to the display of the popup for TokenDropdown. When opening it, a rendering error occurred that prevented the view from being displayed. The reason was that the static width for this widget was removed on this branch, but it was required in this case.

List of changes:
- added static width for popup button widget in token_dropdown.dart to fix render error during its opening