Feature: Infra cache (#65)
This feature introduces a caching system for all API queries. Previously, the application performed a new request to the server when any page was opened or reopened. This caused that navigation within the application was time-consuming, especially for views with slow endpoints such as QueryTransactions. For this reason, each request currently is saved in cache and stored for 1 minute by default (this value can be adjusted for each endpoint). During this time, if the application executes a query, a response from cache will be returned instead of querying server. In addition, each request can be force-refreshed allowing query server regardless of the cache value.

List of changes:
- created http_client_interceptor.dart to intercept API query and choose whether to return a response from the API or from the cache
- created api_cache_repository.dart and api_cache_manager.dart to manage CRUD operations on cached API responses
- created api_cache_response_entity.dart to store database structure for cached API responses
- created api_cache_response_model.dart - business object used within application to represent cached response
- created api_cache_config_model.dart - business model containing cache configuration for specified endpoints
- created new “defaultApiCacheAge” parameter in app_config.dart representing default cache max-age for each endpoint
- created new “withMockedInitialValues()” constructor in memory_cache_manager.dart to simplify tests for classes using it
- added new “api_cache” key to Hive initialization (hive.dart) to allow storing API responses in database
- added headers to mocked responses in mock_api_repository.dart and mock_api_kira_repository.dart
- added new dev dependency - “http_mock_adapter”. This package is used to mock API responses and their statuses. It was required to create tests for http_client_interceptor.dart.