Feature: My Account - Staking tab UI
This feature introduces new tab containing list of staking pools, to which tokens were delegated. This is the base section for Delegation, Undelegation and Claim Rewards transactions, that will be implemented afterwards. This section will allow delegating and undelegating tokens to specific validators and to claim the rewards generated in the network.

List of changes:
- added Staking tab in My Account Page
- created new drawer page accessed from My Account - Staking tab, which provides staking pool details of selected Validator
- created staking_pool_drawer_cubit.dart to manage the states of staking pool data in the drawer
- created account_tile_copy_wrapper.dart as a generic widget to unify the views containing AccountTile with copy button and info tip
- modified query_delegations_service.dart, to make it compatible with changes introduced by infra cache feature
- unrelated with the branch's specific domain: added ICacheManager initialization in test_utils.dart to make integration tests compatible with changes introduced by infra cache feature