Feature: Delegate/Stake Tokens Tx UI (#73)
This feature introduces new page containing transaction process for delegating (staking) tokens to a specific validator. It allows selecting the token, inputting its value, then signing the transaction and sending this data to the blockchain to delegate this token to selected validator. Because of work organization reasons, this branch does not yet support delegating multiple tokens.

List of changes:
- created new page staking_tx_delegate_page.dart containing implementation of transaction process for MsgDelegate
- created form widgets for MsgDelegate (staking_msg_delegate_form.dart, staking_msg_delegate_form_preview.dart)
- added buttons to initialize MsgDelegate transaction in sections: Validator Drawer Page and My Account - Staking tab
- created tx_validator_preview.dart containing a lighter, non-editable version of widget from wallet_address_text_field.dart, which displays username instead of address