Bugfix: Default lowest denomination
The main goal for this branch is to set the lowest denomination as default and swap places of token denomination in transaction view. Previous order was from the highest to the lowest. After change, it is from the lowest to the highest. The additional goal for this branch is to globally refactor names of denomination models: defaultTokenDenomination -> networkTokenDenomination and lowestTokenDenomination -> defaultTokenDenomination.

List of changes:
- changed getter method with order of token denomination models to swap from lowest to highest in token_alias_model.dart
- changed factory return in token_form_state.dart from default token denomination to the lowest token denomination
- changed default token denomination model in token_form_cubit.dart and token_form_state.dart to the lowest token denomination model
- updated token amount preview widget, to use calculated token value based on selected denomination. Previously this widget contained simple getters, which didn't update widget when user changed denomination.
- renamed defaultTokenDenomination -> networkTokenDenomination and lowestTokenDenomination -> defaultTokenDenomination throughout the application