Bugfix: Last block time (#77)
The first goal of this branch is to fix the pixel overflow on last block time widget, that appeared because recently added "refresh" button took some of its space. It made impossible to display full widget in mobile view in case of outdated block time. The second goal is to restore the LastBlockTimeWidget in My Account - IR tab, that was deleted after implementing Infra Cache UI (version 1.13.4).

List of changes:
- modified last_block_time_widget.dart to split the text in two lines in mobile view
- changed height parameter in sliver_list_layout.dart to ensure wrapped text has enough space at the bottom
- restored LastBlockTimeWidget in My Account - Identity Registrar tab from version 1.13.4
- created block_time_wrapper_model.dart to wrap response models with block time date. It works similar to PageData, but is focused on single models, not model lists.
- changed last block time date format to "DD MMM YYYY, HH:mm" (ex. 20 Dec 2023, 09:25)