Bugfix: Loading Page navigation
The purpose of this branch was to examine and fix the bug related to Loading Page navigation. It could occur at the start of the application, after clicking "Cancel connection" button and connecting to a network. In case of connecting to network from assets, there was an unexpected flipping view and the Value Notifier Exception was thrown. In case of connecting to custom network (but only unhealthy), the app navigated to DashboardPage with no network connected.

List of changes:
 - removed emit from "cancelConnection()" method in loading_page_cubit.dart. It was redundant, because adding NetworkModuleDisconnectEvent causes "_handleNetworkModuleStateChanged()", which then emits proper LoadingPageState - two states emitted in very short time caused flipping view. Also, the removed emit did not set value for NetworkModuleState, which caused "no network error".
 - modified "_mapAutoConnectEventToState()" method network_module_bloc.dart to prevent the "disconnected" status from being overwritten, in case of clicking "Cancel connection" button