Bugfix: Favourite balance refresh
The purpose of this branch is to fix not refreshing favourite balance right after sending tokens. Additionally, it was noticed that on desktop view, adding balances to favourites is broken if there is one record, because emitting list with changed favourites by InfinityListBloc does not cause SliverInfinityList to reload. So this issue was also examined on this branch.

List of changes:
- added missing "forceRequestBool" parameter while calling initFavourites() in _mapListReloadEventToState() method in a_list_bloc.dart. It allowed refreshing balances by avoiding getting them from cache.
- added emit(ListLoadingState()) in _mapListUpdatedEventToState() method in infinity_list_bloc.dart, to ensure the SliverInfinityList is reloaded with each emit, and the favourites will be updated. However, this is not the optimal solution, it may be necessary to rebuild the lists to manage this issue properly. Details can be found in "unify-list_items" ticket on Trello.