Feature: MetaMask login
Implemented MetaMask login functionality consists of connection, handling callbacks, including tests for most of the logic. User has now also ability to log in using ETH account of MetaMask or using web wallet(old ways: Keyfile and Mnemonic). After user tap on MetaMask login, it suggests to switch the chain to KIRA's with the according rpc url (for now it's fake). During the session app is listening for changes of the MetaMask's accounts and re-logs in.

List of changes:
- added metamask_cubit.dart with the state to maintain the UI logic of MetaMask functionality
- added flutter_web3 package for managing the MetaMask ETH requests
- wrote metamask_cubit_test.dart to test most of the cases using blocTest and mockito for mocking of the dependencies
- fixed encoding of ethereum_wallet_address.dart to the `HexCodec` as it's appropriate for ETH address operations in bytes
- swapped fake KIRA's address inside identity_records_service.dart to mapped KIRA address from ETH address in case of MetaMask's session
- fixed assertion check for [lastDerivationPathSegment] inside wallet.dart