Feature: Hardcode token aliases instead of fetching INTERX
Hardcoded aliases response with `ukex` as denomination and `kira` as prefix. Made deprecated amount in TokenAlias as its dynamic field in static model, and it cannot be provided from any service except INTERX+SEKAI.
We decided to hardcode aliases instead of fetching from INTERX / using a Remote Config for that. A remote service is pretty much useless in our case. CDN / other BE service would be too difficult to setup just for this task. It can be hardcoded instead for now, and business will not be affected.
SEKAI should not be responsible for aliases except dynamic denomination, but this thing is postponed for an unknown period of time.

List of changes:
- hardcoded aliases response (QueryKiraTokensAliasesResp) with `ukex` as denomination and `kira` as prefix
- fixed tests of aliases service